two step deletion for the doc comment

+++ languages/baseLanguage/javadoc/source_gen/jetbrains/mps/baseLanguage/javadoc/editor/RemoveDocComment.java
@@ -11,14 +11,15 @@ import jetbrains.mps.internal.collections.runtime.ListSequence;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SNodeOperations;
 import jetbrains.mps.smodel.adapter.structure.MetaAdapterFactory;
 import org.jetbrains.mps.openapi.language.SAbstractConcept;
+import jetbrains.mps.editor.runtime.deletionApprover.DeletionApproverUtil;
 import jetbrains.mps.editor.runtime.selection.SelectionUtil;
 import jetbrains.mps.openapi.editor.selection.SelectionManager;
 
 public class RemoveDocComment {
   public static void setCellActions(EditorCell editorCell, SNode node, EditorContext context) {
     editorCell.setAction(CellActionType.DELETE, new RemoveDocComment.RemoveDocComment_DELETE(node));
-    editorCell.setAction(CellActionType.BACKSPACE, new RemoveDocComment.RemoveDocComment_BACKSPACE(node));
     editorCell.setAction(CellActionType.INSERT, new RemoveDocComment.RemoveDocComment_INSERT(node));
+    editorCell.setAction(CellActionType.BACKSPACE, new RemoveDocComment.RemoveDocComment_BACKSPACE(node));
   }
   public static class RemoveDocComment_DELETE extends AbstractCellAction {
     /*package*/ SNode myNode;
@@ -30,21 +31,9 @@ public class RemoveDocComment {
     }
     public void execute_internal(EditorContext editorContext, SNode node) {
       SNode doc = ListSequence.fromList(SNodeOperations.getNodeDescendants(node, MetaAdapterFactory.getConcept(0xf280165065d5424eL, 0xbb1b463a8781b786L, 0x4a3c146b7fae70d3L, "jetbrains.mps.baseLanguage.javadoc.structure.BaseDocComment"), true, new SAbstractConcept[]{})).first();
-      SNode commentedNode = SNodeOperations.getParent(doc);
-      SNodeOperations.deleteNode(doc);
-      SelectionUtil.selectCell(editorContext, commentedNode, SelectionManager.FIRST_CELL);
-    }
-  }
-  public static class RemoveDocComment_BACKSPACE extends AbstractCellAction {
-    /*package*/ SNode myNode;
-    public RemoveDocComment_BACKSPACE(SNode node) {
-      this.myNode = node;
-    }
-    public void execute(EditorContext editorContext) {
-      this.execute_internal(editorContext, this.myNode);
-    }
-    public void execute_internal(EditorContext editorContext, SNode node) {
-      SNode doc = ListSequence.fromList(SNodeOperations.getNodeDescendants(node, MetaAdapterFactory.getConcept(0xf280165065d5424eL, 0xbb1b463a8781b786L, 0x4a3c146b7fae70d3L, "jetbrains.mps.baseLanguage.javadoc.structure.BaseDocComment"), true, new SAbstractConcept[]{})).first();
+      if (DeletionApproverUtil.approve(editorContext, doc, "commentCollection")) {
+        return;
+      }
       SNode commentedNode = SNodeOperations.getParent(doc);
       SNodeOperations.deleteNode(doc);
       SelectionUtil.selectCell(editorContext, commentedNode, SelectionManager.FIRST_CELL);
@@ -60,6 +49,24 @@ public class RemoveDocComment {
     }
     public void execute_internal(EditorContext editorContext, SNode node) {
       CommentLineEditingUtil.insertLine(editorContext);
+    }
+  }
+  public static class RemoveDocComment_BACKSPACE extends AbstractCellAction {
+    /*package*/ SNode myNode;
+    public RemoveDocComment_BACKSPACE(SNode node) {
+      this.myNode = node;
+    }
+    public void execute(EditorContext editorContext) {
+      this.execute_internal(editorContext, this.myNode);
+    }
+    public void execute_internal(EditorContext editorContext, SNode node) {
+      SNode doc = ListSequence.fromList(SNodeOperations.getNodeDescendants(node, MetaAdapterFactory.getConcept(0xf280165065d5424eL, 0xbb1b463a8781b786L, 0x4a3c146b7fae70d3L, "jetbrains.mps.baseLanguage.javadoc.structure.BaseDocComment"), true, new SAbstractConcept[]{})).first();
+      if (DeletionApproverUtil.approve(editorContext, doc, "commentCollection")) {
+        return;
+      }
+      SNode commentedNode = SNodeOperations.getParent(doc);
+      SNodeOperations.deleteNode(doc);
+      SelectionUtil.selectCell(editorContext, commentedNode, SelectionManager.FIRST_CELL);
     }
   }
 }
